L3/Building your First AI Agent with LangGraph/requirements.txt
# Core AI Framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
langchain-openai>=0.2.0

# Web UI
streamlit>=1.38.0

# Data Models
pydantic>=2.0.0
pydantic-settings>=2.0.0

# HTTP & APIs
httpx>=0.27.0
requests>=2.31.0

# Caching
diskcache>=5.6.0

# Environment
python-dotenv>=1.0.0

# Report Generation
tabulate>=0.9.0
fpdf2>=2.8.0

# Utilities
rich>=13.0.0
